\name{casen}
\alias{casen}
\docType{data}
\title{Casen data}
\description{
  
}
\usage{data(casen)}
\format{
  \describe{
    \name{casen}
\item{segmento}{segmento}
\item{folio}{folio}
\item{o}{o}
\item{region}{region}
\item{comuna}{comuna}
\item{zona}{zona}
\item{estrato}{estrato}
\item{varstrat}{varstrat}
\item{varunit}{varunit}
\item{expr_r2}{expr_r2}
\item{expc_r2}{expc_r2}
\item{fmr2}{fmr2}
\item{pco1}{pco1}
\item{viv}{viv}
\item{hog}{hog}
\item{tot_hogviv}{tot_hogviv}
\item{o_viv}{o_viv}
\item{sexo}{sexo}
\item{edad}{edad}
\item{e1}{e1}
\item{o15}{o15}
\item{o17}{o17}
\item{r6}{r6}
\item{r11a}{r11a}
\item{r11b}{r11b}
\item{r13a}{r13a}
\item{r13b}{r13b}
\item{r13c}{r13c}
\item{r13d}{r13d}
\item{r13e}{r13e}
\item{r13f}{r13f}
\item{esc}{esc}
\item{asiste}{asiste}
\item{activ}{activ}
\item{oficio1}{oficio1}
\item{rama1}{rama1}
\item{corte}{corte}
\item{yoprhaj}{yoprhaj}
\item{ytrabhaj}{ytrabhaj}
\item{yauthaj}{yauthaj}
\item{ymonehaj}{ymonehaj}
\item{ysubhaj}{ysubhaj}
\item{ytothaj}{ytothaj}
\item{yaimhaj}{yaimhaj}
\item{hacinamiento}{hacinamiento}
  }
}
\source{ ?}
\examples{
  data(casen)
}
\keyword{datasets}
